Add missing plugin changes to trunk

+++ src/kits/base/plugins/addhost/02-dnsreverse.py
@@ -19,12 +19,13 @@ from kusu.addhost import *
 
 class AddHostPlugin(AddHostPluginBase):
     def finished(self, nodelist):
-        self.dbconn.execute("SELECT networks.network FROM networks, ng_has_net, nodes WHERE nodes.name='%s' \
-                        AND ng_has_net.netid=networks.netid AND nodes.ngid=ng_has_net.ngid" % nodename)
+        if nodelist:
+           self.dbconn.execute("SELECT networks.network FROM networks, ng_has_net, nodes WHERE nodes.name='%s' \
+                               AND ng_has_net.netid=networks.netid AND nodes.ngid=ng_has_net.ngid" % nodelist[0])
 
-        networks = self.dbconn.fetchall()
+           networks = self.dbconn.fetchall()
 
-        # Run genconfig for each network
-        for net in networks:
-            os.system("/opt/kusu/bin/genconfig reverse %s > /var/named/reverse.%s" % (net[0], net[0]))
-            os.system("kill -HUP `pidof named`")
+           # Run genconfig for each network
+           for net in networks:
+               os.system("/opt/kusu/bin/genconfig reverse %s > /var/named/reverse.%s" % (net[0], net[0]))
+               os.system("kill -HUP `pidof named`")
